+++ CHANGELOG.md
@@ -6,11 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
 ### Fixed
-
 - count() now defaults to '*' only if no limit or offset is set
 - Repository->fetchChunked() is now typehinted too
+- InsertRepositoryEvent now correctly sets Query on the reused entity.
+- PHPStan lvl8
 
 
 ## [0.4.1] - 2024-03-19